requirements.txt
langchain==0.1.10
langchain-openai==0.0.8
python-dotenv==1.0.1
requests==2.31.0
fairscale==0.4.13
torch==2.2.2
accelerate==0.20.3
transformers==4.38.2
gradio==4.31.5
gpt4all==2.0.2
huggingface_hub==0.21.3
pypdf==3.17.3
chromadb==0.4.15
google-generativeai==0.3.2
langchain-google-genai==0.0.9
azure-cognitiveservices-speech==1.38.0
playsound==1.3.0